Correcting a bit of a chicken and egg.
If we didn't rebuild o2r1 would have been difficult to find this problem.
We must first install rabbit with default settings, and start the daemon.
Once rabbit installed and running, we can then setup clustering with a
user/pass that now exists.

Previously it was failing since we created the config file with a user
that did not yet exist.